fix: 중복 로직 수정
