+++ users/tests/views/test_user_view.py
@@ -1,22 +1,16 @@
-import uuid
-from unittest.mock import MagicMock
 
 import pytest
 from django.urls import reverse
 
+from users.tests.conftest import FlexiMock
 from users.exceptions import PasswordMismatchException
-
-
-class FlexiMock(MagicMock):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, spec_set=dict, **kwargs)
-        self.get = lambda x, default=None: getattr(self, x, default)
-        self.__getitem__.side_effect = lambda key: getattr(self, key)
 
 
 @pytest.mark.django_db
 class TestUser:
     def test_profile_get(self, authenticated_client, mocker):
+        user = authenticated_client.user
+
         mock_profile = FlexiMock()
         mock_profile.nickname = "테스터"
 
@@ -47,6 +41,7 @@ class TestUser:
             "users.repositories.user_repository.UserRepository.get_user_profile",
             return_value=mock_profile,
         )
+        # 이 테스트는 Serializer.save가 성공했음을 모킹
         mocker.patch("users.serializers.UserProfileSerializer.save", return_value=None)
 
         url = reverse("user-profile")
@@ -63,10 +58,12 @@ class TestUser:
         response = authenticated_client.patch(url, {"nickname": "newnick"})
         assert response.status_code == 404
 
-    def test_password_change(self, authenticated_client, create_user, mocker):
-        user, pwd = create_user(f"user_{uuid.uuid4().hex}@example.com")
+    def test_password_change(self, authenticated_client, mocker):
+        user = authenticated_client.user
+        pwd = authenticated_client.password
 
         mocker.patch(
+            # authenticated_client가 이미 인증된 상태이므로 user_profile은 user.user_profile에서 접근 가능함
             "users.repositories.user_repository.UserRepository.get_user_profile",
             return_value=user.user_profile,
         )
@@ -75,7 +72,6 @@ class TestUser:
             return_value=True,
         )
 
-        authenticated_client.force_authenticate(user)
         url = reverse("user-password-change")
         response = authenticated_client.patch(
             url,
@@ -87,15 +83,14 @@ class TestUser:
         )
         assert response.status_code == 200
 
-    def test_password_change_invalid(self, authenticated_client, create_user, mocker):
-        user, _ = create_user(f"user_{uuid.uuid4().hex}@example.com")
+    def test_password_change_invalid(self, authenticated_client, mocker):
+        user = authenticated_client.user
 
         mocker.patch(
             "users.services.user_service.UserService.change_user_password",
             side_effect=PasswordMismatchException(),
         )
 
-        authenticated_client.force_authenticate(user)
         url = reverse("user-password-change")
         response = authenticated_client.patch(
             url,
@@ -107,8 +102,9 @@ class TestUser:
         )
         assert response.status_code == 401
 
-    def test_user_delete(self, authenticated_client, create_user, mocker):
-        user, pwd = create_user(f"user_{uuid.uuid4().hex}@example.com")
+    def test_user_delete(self, authenticated_client, mocker):
+        user = authenticated_client.user
+        pwd = authenticated_client.password
 
         mocker.patch(
             "users.repositories.user_repository.UserRepository.get_user_profile",
@@ -119,7 +115,6 @@ class TestUser:
             return_value=True,
         )
 
-        authenticated_client.force_authenticate(user)
         url = reverse("user-delete")
         response = authenticated_client.post(url, {"password": pwd})
         assert response.status_code == 200
@@ -129,17 +124,14 @@ class TestUser:
         response = authenticated_client.post(url, {})
         assert response.status_code == 400
 
-    def test_user_delete_invalid_password(
-        self, authenticated_client, create_user, mocker
-    ):
-        user, _ = create_user(f"user_{uuid.uuid4().hex}@example.com")
+    def test_user_delete_invalid_password(self, authenticated_client, mocker):
+        user = authenticated_client.user
 
         mocker.patch(
             "users.services.user_service.UserService.delete_user",
             side_effect=PasswordMismatchException(),
         )
 
-        authenticated_client.force_authenticate(user)
         url = reverse("user-delete")
         response = authenticated_client.post(url, {"password": "wrongpass"})
         assert response.status_code == 401
